Refactor: use _iter_coroutine() on top level
Also extracted most duplicate code, and dropped outdated code
