+++ edgedb/asyncio_client.py
@@ -20,20 +20,16 @@
 import asyncio
 import functools
 import logging
-import random
 import socket
 import ssl
-import time
 import typing
-import uuid
 
 from . import abstract
 from . import base_client
 from . import compat
 from . import con_utils
-from . import enums
 from . import errors
-from . import retry as _retry
+from . import transaction
 from .protocol import asyncio_proto
 
 
@@ -47,6 +43,7 @@ logger = logging.getLogger(__name__)
 
 class AsyncIOConnection(base_client.BaseConnection):
     __slots__ = ("_loop",)
+    _close_exceptions = (Exception, asyncio.CancelledError)
 
     def __init__(self, loop, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -56,48 +53,14 @@ class AsyncIOConnection(base_client.BaseConnection):
         protocol = self._protocol
         return protocol is None or not protocol.connected
 
-    async def connect(self, *, single_attempt=False):
-        start = time.monotonic()
-        if single_attempt:
-            max_time = 0
-        else:
-            max_time = start + self._config.wait_until_available
-        iteration = 1
+    async def connect_addr(self, addr, timeout):
+        try:
+            await compat.wait_for(self._connect_addr(addr), timeout)
+        except asyncio.TimeoutError as e:
+            raise TimeoutError from e
 
-        while True:
-            for addr in self._addrs:
-                try:
-                    await compat.wait_for(
-                        self._connect_addr(addr),
-                        self._config.connect_timeout,
-                    )
-                except asyncio.TimeoutError as e:
-                    if iteration == 1 or time.monotonic() < max_time:
-                        continue
-                    else:
-                        raise errors.ClientConnectionTimeoutError(
-                            f"connecting to {addr} failed in"
-                            f" {self._config.connect_timeout} sec"
-                        ) from e
-                except errors.ClientConnectionError as e:
-                    if (
-                        e.has_tag(errors.SHOULD_RECONNECT) and
-                        (iteration == 1 or time.monotonic() < max_time)
-                    ):
-                        continue
-                    nice_err = e.__class__(
-                        con_utils.render_client_no_connection_error(
-                            e,
-                            addr,
-                            attempts=iteration,
-                            duration=time.monotonic() - start,
-                        ))
-                    raise nice_err from e.__cause__
-                else:
-                    return
-
-            iteration += 1
-            await asyncio.sleep(0.01 + random.random() * 0.2)
+    async def sleep(self, seconds):
+        await asyncio.sleep(seconds)
 
     def _protocol_factory(self, tls_compat=False):
         return asyncio_proto.AsyncIOProtocol(
@@ -159,106 +122,22 @@ class AsyncIOConnection(base_client.BaseConnection):
         self._protocol = pr
         self._addr = addr
 
-    async def privileged_execute(self, query):
-        await self._protocol.simple_query(query, enums.Capability.ALL)
-
-    async def aclose(self):
-        """Send graceful termination message wait for connection to drop."""
-        if not self.is_closed():
-            try:
-                self._protocol.terminate()
-                await self._protocol.wait_for_disconnect()
-            except (Exception, asyncio.CancelledError):
-                self.terminate()
-                raise
-            finally:
-                self._cleanup()
-
     def _dispatch_log_message(self, msg):
         for cb in self._log_listeners:
             self._loop.call_soon(cb, self, msg)
 
-    def __repr__(self):
-        if self.is_closed():
-            return '<{classname} [closed] {id:#x}>'.format(
-                classname=self.__class__.__name__, id=id(self))
-        else:
-            return '<{classname} [connected to {addr}] {id:#x}>'.format(
-                classname=self.__class__.__name__,
-                addr=self.connected_addr(),
-                id=id(self))
-
-    async def raw_query(self, query_context: abstract.QueryContext):
-        if self.is_closed():
-            await self.connect()
-
-        reconnect = False
-        capabilities = None
-        i = 0
-        while True:
-            i += 1
-            try:
-                if reconnect:
-                    await self.connect(single_attempt=True)
-                return await self._protocol.execute_anonymous(
-                    query=query_context.query.query,
-                    args=query_context.query.args,
-                    kwargs=query_context.query.kwargs,
-                    reg=query_context.cache.codecs_registry,
-                    qc=query_context.cache.query_cache,
-                    io_format=query_context.query_options.io_format,
-                    expect_one=query_context.query_options.expect_one,
-                    required_one=query_context.query_options.required_one,
-                    allow_capabilities=enums.Capability.EXECUTE,
-                )
-            except errors.EdgeDBError as e:
-                if query_context.retry_options is None:
-                    raise
-                if not e.has_tag(errors.SHOULD_RETRY):
-                    raise e
-                if capabilities is None:
-                    cache_item = query_context.cache.query_cache.get(
-                        query=query_context.query.query,
-                        io_format=query_context.query_options.io_format,
-                        implicit_limit=0,
-                        inline_typenames=False,
-                        inline_typeids=False,
-                        expect_one=query_context.query_options.expect_one,
-                    )
-                    if cache_item is not None:
-                        _, _, _, capabilities = cache_item
-                # A query is read-only if it has no capabilities i.e.
-                # capabilities == 0. Read-only queries are safe to retry.
-                # Explicit transaction conflicts as well.
-                if (
-                    capabilities != 0
-                    and not isinstance(e, errors.TransactionConflictError)
-                ):
-                    raise e
-                rule = query_context.retry_options.get_rule_for_exception(e)
-                if i >= rule.attempts:
-                    raise e
-                await asyncio.sleep(rule.backoff(i))
-                reconnect = self.is_closed()
-
-    async def execute(self, query: str) -> None:
-        await self._protocol.simple_query(
-            query, enums.Capability.EXECUTE)
-
 
 class _PoolConnectionHolder(base_client.PoolConnectionHolder):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self._release_event = asyncio.Event()
-        self._release_event.set()
+    __slots__ = ()
+    _event_class = asyncio.Event
 
     async def close(self, *, wait=True):
         if self._con is None:
             return
         if wait:
-            await self._con.aclose()
+            await self._con.close()
         else:
-            self._pool._loop.create_task(self._con.aclose())
+            self._pool._loop.create_task(self._con.close())
 
     async def wait_until_released(self, timeout=None):
         await self._release_event.wait()
@@ -278,20 +157,20 @@ class _AsyncIOPoolImpl(base_client.BasePoolImpl):
         on_release=None,
         connection_class,
     ):
-        super().__init__(
-            connect_args,
-            connection_class=connection_class,
-            concurrency=concurrency,
-            on_connect=on_connect,
-            on_acquire=on_acquire,
-            on_release=on_release,
-        )
-        self._loop = None
         if not issubclass(connection_class, AsyncIOConnection):
             raise TypeError(
                 f'connection_class is expected to be a subclass of '
                 f'edgedb.asyncio_client.AsyncIOConnection, '
                 f'got {connection_class}')
+        self._loop = None
+        super().__init__(
+            connect_args,
+            lambda *args: connection_class(self._loop, *args),
+            concurrency=concurrency,
+            on_connect=on_connect,
+            on_acquire=on_acquire,
+            on_release=on_release,
+        )
 
     def _ensure_initialized(self):
         if self._loop is None:
@@ -302,11 +181,6 @@ class _AsyncIOPoolImpl(base_client.BasePoolImpl):
 
     def _set_queue_maxsize(self, maxsize):
         self._queue._maxsize = maxsize
-
-    async def _new_connection_with_params(self, addr, config, params):
-        con = self._connection_class(self._loop, [addr], config, params)
-        await con.connect()
-        return con
 
     async def _maybe_get_first_connection(self):
         async with self._first_connect_lock:
@@ -322,11 +196,11 @@ class _AsyncIOPoolImpl(base_client.BasePoolImpl):
             # we close it.  A new connection will be created
             # when `acquire` is called again.
             try:
-                # Use `aclose()` to close the connection gracefully.
+                # Use `close()` to close the connection gracefully.
                 # An exception in `setup` isn't necessarily caused
-                # by an IO or a protocol error.  aclose() will
+                # by an IO or a protocol error.  close() will
                 # do the necessary cleanup via _release_on_close().
-                await con.aclose()
+                await con.close()
             finally:
                 raise ex
 
@@ -359,7 +233,7 @@ class _AsyncIOPoolImpl(base_client.BasePoolImpl):
 
         if not isinstance(holder._con, AsyncIOConnection):
             raise errors.InterfaceError(
-                f'AsyncIOPool.release() received invalid connection: '
+                f'release() received invalid connection: '
                 f'{holder._con!r} does not belong to any connection pool'
             )
 
@@ -376,7 +250,7 @@ class _AsyncIOPoolImpl(base_client.BasePoolImpl):
         Wait until all pool connections are released, close them and
         shut down the pool.  If any error (including cancellation) occurs
         in ``close()`` the pool will terminate by calling
-        AsyncIOPool.terminate() .
+        _AsyncIOPoolImpl.terminate() .
 
         It is advisable to use :func:`python:asyncio.wait_for` to set
         a timeout.
@@ -413,13 +287,59 @@ class _AsyncIOPoolImpl(base_client.BasePoolImpl):
 
     def _warn_on_long_close(self):
         logger.warning(
-            'AsyncIOPool.aclose() is taking over 60 seconds to complete. '
+            'AsyncIOClient.aclose() is taking over 60 seconds to complete. '
             'Check if you have any unreleased connections left. '
             'Use asyncio.wait_for() to set a timeout for '
-            'AsyncIOPool.aclose().')
+            'AsyncIOClient.aclose().')
 
 
-class AsyncIOClient(abstract.AsyncIOExecutor, base_client.BaseClient):
+class AsyncIOIteration(transaction.BaseTransaction, abstract.AsyncIOExecutor):
+
+    __slots__ = ("_managed",)
+
+    def __init__(self, retry, client, iteration):
+        super().__init__(retry, client, iteration)
+        self._managed = False
+
+    async def __aenter__(self):
+        if self._managed:
+            raise errors.InterfaceError(
+                'cannot enter context: already in an `async with` block')
+        self._managed = True
+        return self
+
+    async def __aexit__(self, extype, ex, tb):
+        self._managed = False
+        return await self._exit(extype, ex)
+
+    async def _ensure_transaction(self):
+        if not self._managed:
+            raise errors.InterfaceError(
+                "Only managed retriable transactions are supported. "
+                "Use `async with transaction:`"
+            )
+        await super()._ensure_transaction()
+
+
+class AsyncIORetry(transaction.BaseRetry):
+
+    def __aiter__(self):
+        return self
+
+    async def __anext__(self):
+        # Note: when changing this code consider also
+        # updating Retry.__next__.
+        if self._done:
+            raise StopAsyncIteration
+        if self._next_backoff:
+            await asyncio.sleep(self._next_backoff)
+        self._done = True
+        iteration = AsyncIOIteration(self, self._owner, self._iteration)
+        self._iteration += 1
+        return iteration
+
+
+class AsyncIOClient(base_client.BaseClient, abstract.AsyncIOExecutor):
     """A lazy connection pool.
 
     A Client can be used to manage a set of connections to the database.
@@ -431,76 +351,28 @@ class AsyncIOClient(abstract.AsyncIOExecutor, base_client.BaseClient):
     :func:`~edgedb.asyncio_client.create_async_client`.
     """
 
+    __slots__ = ()
     _impl_class = _AsyncIOPoolImpl
 
     async def ensure_connected(self):
         await self._impl.ensure_connected()
         return self
 
-    def _clear_codecs_cache(self):
-        self._impl.codecs_registry.clear_cache()
-
-    def _set_type_codec(
-        self,
-        typeid: uuid.UUID,
-        *,
-        encoder: typing.Callable[[typing.Any], typing.Any],
-        decoder: typing.Callable[[typing.Any], typing.Any],
-        format: str
-    ):
-        self._impl.codecs_registry.set_type_codec(
-            typeid,
-            encoder=encoder,
-            decoder=decoder,
-            format=format,
-        )
-
-    async def _query(self, query_context: abstract.QueryContext):
-        con = await self._impl.acquire()
-        try:
-            result, _ = await con.raw_query(query_context)
-            return result
-        finally:
-            await self._impl.release(con)
-
-    async def execute(self, query: str) -> None:
-        """Execute an EdgeQL command (or commands).
-
-        Example:
-
-        .. code-block:: pycon
-
-            >>> await con.execute('''
-            ...     CREATE TYPE MyType { CREATE PROPERTY a -> int64 };
-            ...     FOR x IN {100, 200, 300} UNION INSERT MyType { a := x };
-            ... ''')
-        """
-        con = await self._impl.acquire()
-        try:
-            await con._protocol.simple_query(
-                query, enums.Capability.EXECUTE)
-        finally:
-            await self._impl.release(con)
-
     async def aclose(self):
         """Attempt to gracefully close all connections in the pool.
 
         Wait until all pool connections are released, close them and
         shut down the pool.  If any error (including cancellation) occurs
-        in ``close()`` the pool will terminate by calling
-        AsyncIOPool.terminate() .
+        in ``aclose()`` the pool will terminate by calling
+        AsyncIOClient.terminate() .
 
         It is advisable to use :func:`python:asyncio.wait_for` to set
         a timeout.
         """
         await self._impl.aclose()
 
-    def terminate(self):
-        """Terminate all connections in the pool."""
-        self._impl.terminate()
-
-    def transaction(self) -> _retry.AsyncIORetry:
-        return _retry.AsyncIORetry(self)
+    def transaction(self) -> AsyncIORetry:
+        return AsyncIORetry(self)
 
     async def expire_connections(self):
         """Expire all currently open connections.
